http_parser: TfwHttpParser.port was changed from short to unsigned int
__builtin_uaddl_overflow() in __parse_ulong() might write larger than 2 bytes
to port before "if" checks the limit. At the moment, thank to data alignment,
we have extra 2 bytes in the union, but this might change and we'll get
a vulnerability.

+++ fw/http_parser.h
@@ -123,7 +123,7 @@ typedef struct {
 		} date;
 		struct {
 			unsigned int   flags;
-			unsigned short port;
+			unsigned int port;
 		};
 	};
 	union {
